# test: perf deltafor