… src/v/utils/delta_for.h
… src/v/utils/tests/delta_for_characterization_data.cc