perf before
```
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          24
random seed:              2087347245

test                                      iterations      median         mad         min         max      allocs       tasks        inst
cstore_bench.xor_frame_append                  28049    59.846ns     0.161ns    59.559ns    60.012ns       0.000       0.000         0.0
cstore_bench.xor_frame_append_tx               29069   130.213ns     0.500ns   129.237ns   138.863ns      10.000       0.000         0.0
cstore_bench.xor_column_append                 28730    61.279ns     0.198ns    60.908ns    61.588ns       0.000       0.000         0.0
cstore_bench.xor_column_append_tx              26876   128.953ns     1.791ns   125.800ns   138.704ns       6.000       0.000         0.0
cstore_bench.xor_column_append_tx2             27625    32.812ns     0.014ns    32.716ns    32.826ns       1.000       0.000         0.0
cstore_bench.xor_frame_find_4K                 55692    17.937us     4.369ns    17.913us    17.943us       5.000       0.000         0.0
cstore_bench.xor_column_find_4K                45050    22.180us    20.325ns    22.160us    22.277us      16.000       0.000         0.0
cstore_bench.xor_frame_at_4K                   54357    18.350us    24.513ns    18.261us    18.425us       5.000       0.000         0.0
cstore_bench.xor_column_at_4K                 210966     4.698us     1.776ns     4.680us     4.700us       4.000       0.000         0.0
cstore_bench.xor_column_at_4M                  55404    18.013us     5.407ns    17.985us    18.103us       4.000       0.000         0.0
cstore_bench.delta_frame_append                38541    35.414ns     0.057ns    35.304ns    35.499ns       0.000       0.000         0.0
cstore_bench.delta_frame_append_tx             38193   106.884ns     0.235ns   106.465ns   107.119ns      10.000       0.000         0.0
cstore_bench.delta_column_append               37579    35.981ns     0.048ns    35.891ns    36.118ns       0.000       0.000         0.0
cstore_bench.delta_column_append_tx            36989    91.225ns     0.142ns    90.636ns    91.367ns       6.000       0.000         0.0
cstore_bench.delta_column_append_tx2           37187    33.384ns     0.149ns    33.206ns    33.723ns       1.000       0.000         0.0
cstore_bench.delta_frame_find_4K              133627     7.450us     0.434ns     7.449us     7.452us       5.000       0.000         0.0
cstore_bench.delta_column_find_4K             443596     2.221us     0.252ns     2.220us     2.222us       4.000       0.000         0.0
cstore_bench.delta_column_find_4M              60726    16.305us     2.961ns    16.298us    16.316us       4.000       0.000         0.0
cstore_bench.delta_frame_at_4K                130493     7.631us     3.836ns     7.621us     7.635us       5.000       0.000         0.0
cstore_bench.delta_column_at_4K               498739     1.968us     2.713ns     1.965us     1.970us       4.000       0.000         0.0
cstore_bench.delta_column_at_4M                61182    16.326us     4.914ns    16.320us    16.356us       4.000       0.000         0.0
cstore_bench.xor_frame_at_with_index_4K        24716   295.817ns     0.203ns   295.614ns   296.398ns       9.000       0.000         0.0
cstore_bench.xor_column_at_with_index_4K       24039   298.998ns     0.314ns   298.468ns   301.102ns       6.000       0.000         0.0
cstore_bench.column_store_append_baseline        1620    48.851ns     0.047ns    48.416ns    48.981ns       0.000       0.000         0.0
cstore_bench.column_store_append_result          241    74.149ns     0.178ns    73.971ns    75.104ns       0.000       0.000         0.0
cstore_bench.column_store_scan_baseline         1566    15.380us    10.556ns    15.362us    15.403us       0.000       0.000         0.0
cstore_bench.column_store_scan_result            204   714.058us     1.456us   712.602us   716.390us    4820.726       0.000         0.0
cstore_bench.column_store_find_baseline         1577    46.159ns     0.216ns    44.824ns    46.393ns       0.000       0.000         0.0
cstore_bench.column_store_find_result            237    54.926us    77.494ns    54.765us    55.060us    2648.702       0.000         0.0
cstore_bench.column_store_lower_bound_baseline        1602    45.273ns     0.149ns    45.124ns    45.918ns       0.000       0.000         0.0
cstore_bench.column_store_lower_bound_result         237    54.844us    28.688ns    54.690us    54.880us    2648.669       0.000         0.0
cstore_bench.column_store_upper_bound_baseline        1616    54.659ns     0.444ns    54.042ns    55.226ns       0.000       0.000         0.0
cstore_bench.column_store_upper_bound_result         237    48.813us    56.456ns    48.715us    48.998us    2468.494       0.000         0.0
cstore_bench.column_store_last_segment_baseline        1615    41.916ns     0.107ns    41.792ns    43.276ns       0.000       0.000         0.0
cstore_bench.column_store_last_segment_result         241   128.469ns     1.320ns   126.481ns   130.668ns       0.000       0.000         0.0
cstore_bench.column_store_serialize_baseline        1439   103.637us    44.779ns   103.587us   103.702us      40.000       0.000         0.0
cstore_bench.column_store_serialize_result         232   143.621us   165.901ns   143.455us   143.911us    6826.760       0.000         0.0
cstore_bench.column_store_deserialize_baseline         978   287.248us    60.653ns   287.105us   287.399us    2507.000       0.000         0.0
cstore_bench.column_store_deserialize_result         230    76.250us    85.591ns    76.022us    76.340us     303.130       0.000         0.0

```

perf after 

```
single run iterations:    0
single run duration:      1.000s
number of runs:           5
number of cores:          24
random seed:              1550901199

test                                      iterations      median         mad         min         max      allocs       tasks        inst
cstore_bench.xor_frame_append                  35116    41.467ns     0.057ns    41.333ns    41.634ns       0.000       0.000         0.0
cstore_bench.xor_frame_append_tx               34986   116.538ns     0.131ns   116.406ns   117.642ns      10.000       0.000         0.0
cstore_bench.xor_column_append                 31589    43.051ns     0.018ns    43.000ns    43.096ns       0.000       0.000         0.0
cstore_bench.xor_column_append_tx              33077   101.486ns     0.020ns   101.173ns   101.526ns       6.000       0.000         0.0
cstore_bench.xor_column_append_tx2             32768    30.762ns     0.017ns    30.731ns    30.853ns       1.000       0.000         0.0
cstore_bench.xor_frame_find_4K                140757     7.049us    21.704ns     6.985us     7.185us       5.000       0.000         0.0
cstore_bench.xor_column_find_4K                96931    10.261us     7.671ns    10.243us    10.269us      16.000       0.000         0.0
cstore_bench.xor_frame_at_4K                  142038     7.021us    11.570ns     6.998us     7.057us       5.000       0.000         0.0
cstore_bench.xor_column_at_4K                 528429     1.857us     3.980ns     1.829us     1.861us       4.000       0.000         0.0
cstore_bench.xor_column_at_4M                  64751    15.403us    10.442ns    15.387us    15.442us       4.000       0.000         0.0
cstore_bench.delta_frame_append                35825    41.053ns     0.049ns    40.933ns    41.200ns       0.000       0.000         0.0
cstore_bench.delta_frame_append_tx             34829   115.584ns     0.195ns   115.166ns   116.170ns      10.000       0.000         0.0
cstore_bench.delta_column_append               34602    41.057ns     0.039ns    40.940ns    41.171ns       0.000       0.000         0.0
cstore_bench.delta_column_append_tx            31971    98.698ns     0.404ns    98.230ns    99.102ns       6.000       0.000         0.0
cstore_bench.delta_column_append_tx2           32698    32.383ns     0.015ns    32.364ns    32.543ns       1.000       0.000         0.0
cstore_bench.delta_frame_find_4K              166816     5.984us     0.399ns     5.984us     5.997us       5.000       0.000         0.0
cstore_bench.delta_column_find_4K             392498     2.515us     2.310ns     2.508us     2.519us       4.000       0.000         0.0
cstore_bench.delta_column_find_4M              62095    16.046us    46.878ns    15.999us    16.128us       4.000       0.000         0.0
cstore_bench.delta_frame_at_4K                167217     6.000us    29.853ns     5.930us     6.031us       5.000       0.000         0.0
cstore_bench.delta_column_at_4K               601743     1.649us    12.552ns     1.632us     1.662us       4.000       0.000         0.0
cstore_bench.delta_column_at_4M                61948    16.069us     2.441ns    16.066us    16.094us       4.000       0.000         0.0
cstore_bench.xor_frame_at_with_index_4K        27124   193.195ns     0.476ns   191.746ns   193.857ns       9.000       0.000         0.0
cstore_bench.xor_column_at_with_index_4K       26397   197.869ns     0.105ns   197.264ns   198.111ns       6.000       0.000         0.0
cstore_bench.column_store_append_baseline        1621    48.968ns     0.062ns    48.788ns    49.030ns       0.000       0.000         0.0
cstore_bench.column_store_append_result          238    70.118ns     1.097ns    67.231ns    73.626ns       0.000       0.000         0.0
cstore_bench.column_store_scan_baseline         1575    15.351us    24.478ns    15.326us    15.425us       0.000       0.000         0.0
cstore_bench.column_store_scan_result            211   522.780us   218.806ns   522.561us   523.966us    4820.641       0.000         0.0
cstore_bench.column_store_find_baseline         1571    48.663ns     2.478ns    46.185ns    59.807ns       0.000       0.000         0.0
cstore_bench.column_store_find_result            234    52.585us    75.581ns    52.410us    52.661us    2648.722       0.000         0.0
cstore_bench.column_store_lower_bound_baseline        1604    42.748ns     0.028ns    42.227ns    42.776ns       0.000       0.000         0.0
cstore_bench.column_store_lower_bound_result         234    52.515us   106.308ns    52.335us    52.727us    2648.733       0.000         0.0
cstore_bench.column_store_upper_bound_baseline        1619    54.389ns     0.567ns    53.253ns    54.956ns       0.000       0.000         0.0
cstore_bench.column_store_upper_bound_result         235    47.294us    38.885ns    47.255us    47.514us    2468.808       0.000         0.0
cstore_bench.column_store_last_segment_baseline        1620    38.677ns     2.543ns    36.135ns    50.856ns       0.000       0.000         0.0
cstore_bench.column_store_last_segment_result         239   135.950ns     0.109ns   134.992ns   136.477ns       0.000       0.000         0.0
cstore_bench.column_store_serialize_baseline        1440   104.365us    10.960ns   104.354us   104.515us      40.000       0.000         0.0
cstore_bench.column_store_serialize_result         229   145.968us    50.882ns   145.654us   146.325us    6826.687       0.000         0.0
cstore_bench.column_store_deserialize_baseline         980   288.129us   124.189ns   287.903us   288.253us    2507.000       0.000         0.0
cstore_bench.column_store_deserialize_result         228    76.392us   112.596ns    76.208us    76.505us     303.126       0.000         0.0
```

<!--
See https://github.com/redpanda-data/redpanda/blob/dev/CONTRIBUTING.md##pull-request-body
for more details and examples of what is expected in a PR body.

Content in this top section is REQUIRED. Describe, in plain language, the motivation
behind the change (bug fix, feature, improvement) in this PR and how the included
commits address it.

Add the GitHub keyword `Fixes` to link to bug(s) this PR will fix, e.g.
  Fixes #ISSUE-NUMBER, Fixes #ISSUE-NUMBER, ...

If this PR is a backport, link to the original with `Backport of PR`, e.g.
  Backport of PR #PR-NUMBER
-->

## Backports Required

<!-- Checking at least one of the checkboxes is REQUIRED if this PR is not a backport. -->

- [ ] none - not a bug fix
- [ ] none - this is a backport
- [ ] none - issue does not exist in previous branches
- [ ] none - papercut/not impactful enough to backport
- [ ] v23.2.x
- [ ] v23.1.x
- [ ] v22.3.x

## Release Notes

<!--
If the changes in this PR do not need to be mentioned in the release
notes, then don't add a sub-section and simply list `none`, e.g.

* none

Otherwise, adding a sub-section or `none` is REQUIRED if the PR is not a backport PR.
If this is a backport PR, adding contents to this section will override
the release notes section inherited from the original PR to dev.

Add one or more of the sub-sections with a short description bullet
point of the change, e.g.

### Bug Fixes

* Short description of the bug fix if this is a PR to `dev` branch.

### Features

* Short description of the feature. Explain how to configure.

### Improvements

* Short description of how this PR improves existing behavior.

-->