@andijcr 2023-07-27
__uint128_t does not work well with std::unsigned_integral concept in build-oss , will have to relax the requirement