@andijcr 2023-07-31
force push: 
- cleanups of comments
- changed the way unsigned-decomposition is represented. was pairs of (sizeof(uintN_t), bytes_to_skip), now is pairs of (sizeof(uintN_t), **bits_to_skip**)
- characterization data generation fix: ensured that after xor N_BITS is 64 (before it was by chance)

https://buildkite.com/redpanda/redpanda/builds/34138#01899dfc-3f4d-45cd-a2ff-cf9df93c4a9f
failure is in unit test shutdown https://github.com/redpanda-data/redpanda/issues/12509