@andijcr 2023-08-02
force push: moved tests to utils/test, added check that generated characterization data will in fact test for `N_BITS ∈ [ 0, 64]`