@andijcr 2023-08-04
issues are https://github.com/redpanda-data/redpanda/issues/11362 and https://github.com/redpanda-data/redpanda/issues/12605 (new one, unrelated to this pr)