@andijcr 2023-08-23
failures: 
The summary job failed with 
```
docker: Error response from daemon: Head 
"https://registry-1.docker.io/v2/library/alpine/manifests/3": received 
unexpected HTTP status: 503 Service Unavailable.

```
while the build job failed in the end with

```
docker: Error response from daemon: Get 
"https://registry-1.docker.io/v2/": net/http: request canceled 
(Client.Timeout exceeded while awaiting headers).

``` 

otherwise, unit tests and ducktape tests (release/debug) are all green 

 